+++ homework/llist.h
@@ -6,7 +6,7 @@ typedef enum _TLlistStatus {
 typedef struct _TNode TNode;
 typedef TNode * TLlist;
 
-TStatus llistInsert(TLlist *list, unsigned int value);
+TStatus llistInsert(TLlist *list, unsigned long value);
 TStatus llistDestroy(TLlist *list);
 void llistPrint(TLlist p);
 void testPool();
